% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aom.dss.mouse_ADM.R
\docType{data}
\name{ADM.CountsERCC}
\alias{ADM.CountsERCC}
\title{Counts, ERCC, and metadata from the AOM/DSS mouse.}
\format{Matrix containing ERCC counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
ADM.CountsERCC
}
\description{
Counts, ERCC, and metadata from the AOM/DSS mouse.
}
\examples{
data(ADM.CountsERCC)

}
\keyword{datasets}
