% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aom.dss.mouse_ADM.R
\docType{data}
\name{ADM.Meta}
\alias{ADM.Meta}
\title{Counts, ERCC, and metadata from the AOM/DSS mouse.}
\format{Tibble with:}
\usage{
ADM.Meta
}
\description{
Counts, ERCC, and metadata from the AOM/DSS mouse.
}
\examples{
data(ADM.Meta)

}
\keyword{datasets}
