% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_general.R
\name{annotateCellNumber}
\alias{annotateCellNumber}
\title{Annotate cell numbers.}
\arguments{
\item{data}{tibble; A tibble containing the plate, row, and column
designation for the samples to be annotated.}

\item{plate}{character; A character string indicating the plate variable
where samples to be annotated are located.}

\item{row}{numeric; A numeric vector indicating the rows that are singlets.}

\item{column}{numeric; A numeric vector indicating the columns that are
singlets.}
}
\value{
The metadata tibble with an additional column, named "cellNumber",
 indicating singlets and multiplets via a character of either "Singlet" or
 "Multiplet".
}
\description{
Annotates if cells are singlets or multiplets.
}
\author{
Jason Serviss
}
