% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_countsMgfp.R
\name{annotateGFP}
\alias{annotateGFP}
\title{Annotate sample GFP status.}
\arguments{
\item{data}{tibble; A tibble containing the plate, row, and column
designation for the samples to be annotated.}

\item{plate}{list; A list where each element is a character vector indicating
the plate where samples to be annotated are located.}

\item{row}{list; A list where each element is a numeric vector indicating the
rows that are GFP positive in the corresponding plate.}

\item{column}{list; A list where each element is a numeric vector indicating
the columns that are GFP positive in the corresponding plate.}
}
\value{
The metadata tibble with an additional column indicating GFP staus.
}
\description{
Annotates a logical value indicating a sample's GFP status.
}
\author{
Jason Serviss
}
