% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_countsMgfp.R
\name{annotateMouse}
\alias{annotateMouse}
\title{Annotate mouse.}
\arguments{
\item{data}{tibble; A tibble containing the plate designation for the samples
to be annotated.}

\item{plate}{character; A character vector indicating the plate where samples
to be annotated are located.}

\item{mouseID}{character; ID of the mouse for the corresponding plate.}
}
\value{
The metadata tibble with an additional column indicating mouse ID in
 a column named "mouseID".
}
\description{
Annotates the mouse ID for the experiment. For the moment, assumes that each
plate contains only one mouse.
}
\author{
Jason Serviss
}
