% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_countsMgfp.R
\name{annotateTissue}
\alias{annotateTissue}
\title{Annotate tissue.}
\arguments{
\item{data}{tibble; A tibble containing the plate designation for the samples
to be annotated.}

\item{plate}{character; A character vector indicating the plate where samples
to be annotated are located.}

\item{tissue}{character; Tissue for the corresponding plate.}
}
\value{
The metadata tibble with an additional column indicating tissue in
 a column named "tissue".
}
\description{
Annotates the tissue for the experiment. For the moment, assumes that each
plate contains only one tissue.
}
\author{
Jason Serviss
}
