% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aomdssMouse.R
\docType{data}
\name{countsAomdss}
\alias{countsAomdss}
\title{Counts, ERCC, and metadata from the AOM/DSS mouse.}
\format{Matrix containing counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
countsAomdss
}
\description{
Counts, ERCC, and metadata from the AOM/DSS mouse.
}
\examples{
data(countsAomdss)

}
\keyword{datasets}
