% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aomdssMouse.R
\docType{data}
\name{countsAomdssERCC}
\alias{countsAomdssERCC}
\title{Counts, ERCC, and metadata from the AOM/DSS mouse.}
\format{Matrix containing ERCC counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
countsAomdssERCC
}
\description{
Counts, ERCC, and metadata from the AOM/DSS mouse.
}
\examples{
data(countsAomdssERCC)

}
\keyword{datasets}
