% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aomdssMouse.R
\docType{data}
\name{countsAomdssMeta}
\alias{countsAomdssMeta}
\title{Counts, ERCC, and metadata from the AOM/DSS mouse.}
\format{Tibble with:}
\usage{
countsAomdssMeta
}
\description{
Counts, ERCC, and metadata from the AOM/DSS mouse.
}
\examples{
data(countsAomdssMeta)

}
\keyword{datasets}
