% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetalPancreasCounts.R
\docType{data}
\name{countsErccFp}
\alias{countsErccFp}
\title{Fetal pancreas counts and ERCC reads data.}
\format{matrix
\describe{
    \item{rownames}{Gene names. ERCC identified with "^ERCC-[0-9]*"}
    \item{colnames}{Samples. Singlets prefixed with "s" and multiplets "m".}
}}
\value{
Matrix of counts.
}
\description{
Fetal pancreas counts and ERCC data.
}
\examples{
data(countsErccFp)
}
