% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsMgfp.R
\docType{data}
\name{countsMgfp}
\alias{countsMgfp}
\title{Counts data from the GFP-Lgr5 mouse small intestine and colon.}
\format{Matrix containing counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
countsMgfp
}
\description{
Counts data from the GFP-Lgr5 mouse small intestine and colon.
}
\examples{
data(countsMgfp)

}
\keyword{datasets}
