% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsMgfp.R
\docType{data}
\name{countsMgfpERCC}
\alias{countsMgfpERCC}
\title{ERCC counts data from the GFP-Lgr5 mouse small intestine and colon.}
\format{Matrix containing ERCC counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
countsMgfpERCC
}
\description{
Counts and ERCC data from the GFP-Lgr5 mouse small intestine and colon.
}
\examples{
data(countsMgfpERCC)

}
\keyword{datasets}
