% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsMgfp.R
\docType{data}
\name{countsMgfpMeta}
\alias{countsMgfpMeta}
\title{Metadata from the GFP-Lgr5 mouse small intestine and colon.}
\format{Tibble with:
\describe{
\item{sample}{Sample ID}
\item{plate}{Plate ID}
\item{row}{Row position in plate}
\item{column}{Column position in plate}
\item{GFP}{Samples GFP status}
\item{mouseID}{mouse ID}
\item{tissue}{tissue from which the sample was derived}
\item{cellNumber}{Indicates if sample is a single cell or multiple cells}
}}
\usage{
countsMgfpMeta
}
\description{
Counts, ERCC, and metadata from the GFP-Lgr5 mouse small intestine and colon.
}
\examples{
data(countsMgfpMeta)

}
\keyword{datasets}
