% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsRegev.R
\docType{data}
\name{countsRegev}
\alias{countsRegev}
\title{Counts data from GSE92332 for mouse small intestine.}
\format{Matrix containing counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE92332}
}
\usage{
countsRegev
}
\description{
Counts data from GSE92332 for mouse small intestine.
}
\examples{
data(countsRegev)

}
\keyword{datasets}
