% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsSorted2.R
\docType{data}
\name{countsSorted2}
\alias{countsSorted2}
\title{Counts and ERCC matrix for sorted cell lines.}
\format{Two matrices, containing counts and ERCC counts, with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
countsSorted2 or countsSortedERCC2
}
\value{
matrix
}
\description{
The HOS, HCT116, and A375 cell lines were FACS sorted, either as
singlets or multiplets, and sequenced. Multiplets were sequenced into known
combinations in order to have previousl knowledge regarding which connections
should be contained in each multiplet.
}
\examples{
countsSorted2
countsSortedERCC2

}
