% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_countsSorted2.R
\name{countsSorted2_cellTypes}
\alias{countsSorted2_cellTypes}
\title{Annotate cell types in each well for the countsSorted2 data.}
\arguments{
\item{data}{tibble; A tibble containing the sample names using standard
nomenclature in a column named "sample".}
}
\value{
The metadata tibble with an additional column, named "cellTypes",
 indicating the cell types in the sample. In the case of multiplets, cell
 types are seperated by "-".
}
\description{
Annotate cell types in each well for the countsSorted2 data.
}
\author{
Jason Serviss
}
