% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsSorted1.R
\docType{data}
\name{countsSortedERCC1}
\alias{countsSortedERCC1}
\title{ERCC counts data for sorted multiplets; experiment 1.}
\format{Matrix containing ERCC counts, with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
data(countsSortedERCC1)
}
\description{
Counts and ERCC matrix for sorted cell lines.
}
\details{
The HOS, HCT116, and A375 cell lines were FACS sorted, either as
singlets or multiplets, and sequenced. Multiplets were sequenced into known
combinations in order to know which connections are contained in each
multiplet. Note that many of the ERCC read fractions did not look good for
many of these cells and the sort and sequencing was repeated. The results
from the repeated experiment are in the countsSorted2 dataset.
}
\examples{
data(countsSortedERCC1)

}
\keyword{datasets}
