% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsSorted2.R
\docType{data}
\name{countsSortedERCC2}
\alias{countsSortedERCC2}
\title{ERCC counts data from sorted multiplets; experiment 2.}
\format{Matrix containing ERCC counts, with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
data(countsSortedERCC2)
}
\description{
Counts and ERCC matrix for sorted cell lines.
}
\details{
The HOS, HCT116, and A375 cell lines were FACS sorted, either as
singlets or multiplets, and sequenced. Multiplets were sequenced into known
combinations in order to know which connections are contained in each
multiplet. This dataset is used in conjunction with the
inst/sortedMultiplets20171116/*.Rmd analyses in the sp.scRNAseqTesting
package.
}
\examples{
data(countsSortedERCC2)

}
\keyword{datasets}
