% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsSorted2.R
\docType{data}
\name{countsSortedMeta2}
\alias{countsSortedMeta2}
\title{Counts data from sorted multiplets; experiment 2.}
\format{Two matrices, containing counts and ERCC counts, with:
\describe{
\item{sample}{sample ID}
\item{plate}{plate ID}
\item{row}{Row position in plate}
\item{column}{Column position in plate}
\item{cellNumber}{Indicates if sample is a single cell or multiple cells}
\item{cellTypes}{Indicates samples cell types}
}}
\usage{
countsSortedMeta2
}
\description{
Counts, ERCC, and metadata for sorted cell lines.
}
\details{
The HOS, HCT116, and A375 cell lines were FACS sorted, either as
singlets or multiplets, and sequenced (171116). Multiplets were sequenced
into known combinations in order to know which connections are contained in
each multiplet. This dataset is used in conjunction with the
inst/sortedMultiplets20171116/*.Rmd analyses in the sp.scRNAseqTesting
package.
}
\examples{
data(countsSortedMeta2)
}
\keyword{datasets}
