% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFuns.R
\name{detectLowQualityCells.totalCounts}
\alias{detectLowQualityCells.totalCounts}
\title{detectLowQualityCells.totalCounts}
\arguments{
\item{counts}{data.frame; A data frame with counts data with gene names as
rownames and sample names as colnames.}

\item{mincount}{numeric; A minimum colSum for which columns with a higher
colSum will be detected. Default = 4e5.}
}
\value{
A logical vector with length = ncol(counts) that is TRUE when the
counts data.frame column contains a sample with colSums > mincount.
}
\description{
It is often the case that some samples from sequencing experiments are of
low quality, in many cases due to issues during the sample preperation stage.
Due to the fact that these samples represent a high level of technical noise,
it is often desirable to remove these before downstream analysis which is
facilitated by this function. The function achieves this by detecting cells
whose sum across all genes is > mincount.
}
\examples{
set.seed(8292)
x <- runif(2e4)
y <- runif(2e4, 1, 100)
names <- paste0(letters, 1:2e4)
counts <- data.frame(a = x, b = y, c = y, row.names = names)
detectLowQualityCells.totalCounts(counts)

}
\author{
Jason Serviss
}
