% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFuns.R
\name{filterCountsData}
\alias{filterCountsData}
\title{filterCountsData}
\arguments{
\item{counts}{data.frame; The formated counts data.}

\item{countsERCC}{data.frame; The formated counts ERCC data.}

\item{filters}{Character; A vector indicating the types of filtering to be
performed. Options include: "genes", "totalCounts",
"ERCCfrac", and "housekeeping".}

\item{geneMinCount}{See \code{\link{detectLowQualityGenes}}}

\item{cellMinCount}{See \code{\link{detectLowQualityCells.totalCounts}}}

\item{geneName}{See \code{\link{detectLowQualityCells.housekeeping}}}

\item{quantileCut}{See \code{\link{detectLowQualityCells.housekeeping}}}

\item{percentile}{See \code{\link{detectLowQualityCells.ERCCfrac}}.}
}
\value{
A list with the filtered counts as the first element and the filtered
ERCC genes as the second element.
}
\description{
This is a wrapper function for \code{\link{detectLowQualityGenes}},
\code{\link{detectLowQualityCells.totalCounts}},
\code{\link{detectLowQualityCells.ERCCfrac}},
\code{\link{detectLowQualityCells.housekeeping}},
\code{\link{convertCountsToMatrix}}. It uses the
"cellNumber" column in the metadata to annotate singlets and multiplets and,
therefore, the column must be present in the metadata. In addition, singlets
should be named "Singlet" in the metadata. The function also checks for the
presence of NAs in the counts data and returns an error if NA values are
detected.
}
\author{
Jason Serviss
}
