% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFuns.R
\name{formatCountData}
\alias{formatCountData}
\title{formatCountData}
\arguments{
\item{counts}{data.frame; The unfiltered counts data including ERCC genes
typically arising from the \code{\link{getCountsData}} function.}

\item{metadata}{tibble; The metadata typically arising from the
\code{\link{getMetadata}} function.}
}
\value{
A list with the formated counts as the first element and the formated
ERCC reads as the second element.
}
\description{
This is a wrapper function for \code{\link{moveGenesToRownames}},
\code{\link{labelSingletsAndMultiplets}},
\code{\link{detectERCCreads}}, \code{\link{removeHTSEQsuffix}},
and \code{\link{detectNonGenes}}. It uses the "cellNumber" column in the
metadata to annotate singlets and multiplets and, therefore, the column must
be present in the metadata. In addition, singlets should be named "Singlet"
in the metadata. The function also checks for the presence of NAs in the
counts data and returns an error if NA values are detected.
}
\author{
Jason Serviss
}
