% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFuns.R
\name{getCountsData}
\alias{getCountsData}
\title{getCountsData}
\arguments{
\item{projectName}{character; The name of the project. Must correspond with
the project folder name located at EngeLab/data.}
}
\value{
A data.frame with the unfiltered counts data.
}
\description{
Downloads the count data, located at EngeLab/project name/raw_data, for a
project and concatenates it into a data.frame. The column containing the gene
name, to be named HGN, is expected in all counts files.
}
\author{
Jason Serviss
}
