% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMeanPermutations.R
\docType{data}
\name{groupMeanPermutations}
\alias{groupMeanPermutations}
\title{Permutations of the \code{groupMeans} slot of an spUnsupervised object for
countsSorted2 dataset.}
\format{A matrix, containing the mean gene expression for each
group/cell type, with:
\describe{
\item{rownames}{Permutation number}
\item{colnames}{Cell type}
}}
\usage{
groupMeanPermutations
}
\value{
matrix
}
\description{
The countsSorted2 dataset was used to run the spUnsupervised method,
proceeded by the spCounts method, allowing the groupMeans matrix to be
generated. The cell types, which are known see \code{help(countsSorted2},
were renamed in the spUnsupervised object and the groupMeans slot was
extracted. This was then permuted in such a fashion that, for each row, the
data was reshuffled. This was performed 8100 times which is equal to 100
permutations per multiplet in the dataset. Only the 2000 most highly
expressed genes were used in the spUnsupervised method and, thus, each row in
the groupMeanPermutations corresponds to gene expression values with every
2000 rows corresponding to the next permutation of the set of 2000 genes.
}
\examples{
groupMeanPermutations

}
