% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.Fetal.pancreas_HFP.R
\docType{data}
\name{HFP.Counts}
\alias{HFP.Counts}
\title{Fetal pancreas counts data.}
\format{matrix
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
HFP.Counts
}
\description{
Fetal pancreas counts, ERCC, and meta data.
}
\examples{
data(HFP.Counts)
}
\keyword{datasets}
