% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.Fetal.pancreas_HFP.R
\docType{data}
\name{HFP.CountsERCC}
\alias{HFP.CountsERCC}
\title{Fetal pancreas ERCC data.}
\format{matrix
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
HFP.CountsERCC
}
\description{
Fetal pancreas counts, ERCC, and meta data.
}
\examples{
data(HFP.CountsERCC)
}
\keyword{datasets}
