% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.Fetal.pancreas_HFP.R
\docType{data}
\name{HFP.Meta}
\alias{HFP.Meta}
\title{Fetal pancreas metadata.}
\format{Tibble}
\usage{
HFP.Meta
}
\description{
Fetal pancreas counts, ERCC, and meta data.
}
\examples{
data(HFP.Meta)
}
\keyword{datasets}
