% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFuns.R
\name{labelSingletsAndMultiplets}
\alias{labelSingletsAndMultiplets}
\title{labelSingletsAndMultiplets}
\arguments{
\item{data}{data.frame or character; A data frame with sample IDs as colnames
or a vector of sample IDs.}

\item{ids}{character; a character vector of regex statments used to indicate
colnames of columns that contain SINGLETS.}
}
\value{
If data is of class data.frame colnames are modified in place with
the appropriate prefix attatched. If data is of class character, the modified
character vector is returned. Throws a warning if no singlets are detected.
}
\description{
Adds the prefix "s." to colnames of columns that contain singlets and "m." to
columns that contain multiplets.
}
\examples{

counts <- data.frame(a = runif(26), b = runif(26), c = runif(26))
labelSingletsAndMultiplets(counts, ids = c("a", "b"))

}
\author{
Jason Serviss
}
