% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_general.R
\name{loadMetaData}
\alias{loadMetaData}
\title{Load meta data.}
\arguments{
\item{path}{character; The path to the metadata file. Typically in a
subfolder of inst/rawData.}
}
\value{
The metadata as a tibble.
}
\description{
Loads a metadata .txt file into R.
}
\author{
Jason Serviss
}
