% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse.gut.architecture_MGA.R
\docType{data}
\name{MGA.CountsERCC}
\alias{MGA.CountsERCC}
\title{ERCC counts data from the GFP-Lgr5 mouse small intestine and colon.}
\format{Matrix containing ERCC counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
MGA.CountsERCC
}
\description{
Counts, ERCC, and metadata from the GFP-Lgr5 mouse small intestine and colon.
}
\examples{
data(MGA.CountsERCC)

}
\keyword{datasets}
