% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse.gut.architecture_MGA.R
\docType{data}
\name{MGA.Meta}
\alias{MGA.Meta}
\title{Metadata from the GFP-Lgr5 mouse small intestine and colon.}
\format{Tibble}
\usage{
MGA.Meta
}
\description{
Counts, ERCC, and metadata from the GFP-Lgr5 mouse small intestine and colon.
}
\examples{
data(MGA.Meta)

}
\keyword{datasets}
