% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse.gut.architecture.bulk_MGAB.R
\docType{data}
\name{MGAB.Counts}
\alias{MGAB.Counts}
\title{Counts data from mouse small intestine and colon bulk RNAseq.}
\format{Matrix containing counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples}
}}
\usage{
MGAB.Counts
}
\description{
Counts, ERCC, and metadata from mouse small intestine and colon bulk RNAseq.
}
\examples{
data(MGAB.Counts)

}
\keyword{datasets}
