% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse.gut.architecture.bulk_MGAB.R
\docType{data}
\name{MGAB.CountsERCC}
\alias{MGAB.CountsERCC}
\title{ERCC counts data from mouse small intestine and colon bulk RNAseq.}
\format{Matrix containing ERCC counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples}
}}
\usage{
MGAB.CountsERCC
}
\description{
Counts, ERCC, and metadata from mouse small intestine and colon bulk RNAseq.
}
\examples{
data(MGAB.CountsERCC)

}
\keyword{datasets}
