% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse.gut.architecture.bulk_MGAB.R
\docType{data}
\name{MGAB.Meta}
\alias{MGAB.Meta}
\title{Metadata from mouse small intestine and colon bulk RNAseq.}
\format{Tibble}
\usage{
MGAB.Meta
}
\description{
Counts, ERCC, and metadata from mouse small intestine and colon bulk RNAseq.
}
\examples{
data(MGAB.Meta)

}
\keyword{datasets}
