% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\docType{data}
\name{permutations}
\alias{permutations}
\title{Permutation results from sorted multiplets experiment 2.}
\format{A list containing the permutation results. Can be processed into tidy
format with the \code{tidyPermutationData} function.}
\usage{
data(permutations)
}
\description{
Permutations of the countsSorted2 dataset.
}
\details{
The countsSorted2 dataset was used to run the sp.scRNAseq pipeline. The cell
types, which are known, see \code{help(countsSorted2)}, were renamed in the
spUnsupervised object. The \code{permuteSwarm} was then run and both the
spSwarm object and the output from the \code{permuteSwarm} function were
saved.
}
\examples{
data(permutations)
}
\keyword{datasets}
