% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFuns.R
\name{processedDataUpload}
\alias{processedDataUpload}
\title{processedDataUpload}
\arguments{
\item{projectName}{character; The name of the project. Must correspond with
the project folder name located at EngeLab/data.}

\item{...}{Items to be uploaded as .txt files. Note the files are named such
that they have the projectName argument prepended to the current variable
name.}
}
\value{
The .txt files uploaded to the processed_data folder.
}
\description{
Saves the submitted variables as text files in the drives project folder in
the "processed_data" subfolder. Note that if the "processed_data" folder is
not present in the project dierctory it is created.
}
\author{
Jason Serviss
}
