% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFuns.R
\name{reformatWell}
\alias{reformatWell}
\title{Reformat wells}
\arguments{
\item{names}{character; A character vector of the old names. Names that are
already in the new nomenclature will not be effected.}
}
\value{
A character vector of the updated names.
}
\description{
Reformats wells named e.g. E2 to E02.
}
\author{
Jason Serviss
}
