% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_countsMgfp.R
\name{renameMgfpSamples}
\alias{renameMgfpSamples}
\title{Rename GFP mouse samples.}
\arguments{
\item{oldNames}{character; A character vector of the old names that have
 already been processed by the removeHTSEQsuffix functions. Names that are
already in the new nomenclature will not be effected.}
}
\value{
A character vector of the updated names.
}
\description{
Updates sample names from the old nomenclature (GFPpos.C1.Doublet.5.F9.htseq)
to the new nomenclature (NJA00103.D09.htseq).
}
\author{
Jason Serviss
}
