% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regev.mouse.small.intestine_RSI.R
\docType{data}
\name{RSI.Meta}
\alias{RSI.Meta}
\title{Counts data from GSE92332 for mouse small intestine.}
\format{Tibble}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE92332}
}
\usage{
RSI.Meta
}
\description{
Counts data from GSE92332 for mouse small intestine.
}
\examples{
data(RSI.Meta)

}
\keyword{datasets}
