% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFuns.R
\name{saveRDA}
\alias{saveRDA}
\title{saveRDA}
\arguments{
\item{projectName}{character; The name of the project. Must correspond with
the project folder name located at EngeLab/data.}

\item{...}{Items to be saved in the .rda file. Note that these variables are
renamed before saving such that they have the projectName argument prepended
to the current variable name.}
}
\value{
The .rda file is saved as "data/projectName.rda".
}
\description{
Renames and saves filtered/formated project data as a .rda file in the "data"
folder of the current directory.
}
\author{
Jason Serviss
}
