% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\docType{data}
\name{sObjPermutations}
\alias{sObjPermutations}
\title{spSwarm object with "real" data from permutation analysis.}
\format{An spSwarm object containing the "real" results.}
\usage{
data(sObjPermutations)
}
\description{
Permutations of the countsSorted2 dataset.
}
\details{
The countsSorted2 dataset was used to run the sp.scRNAseq pipeline. The cell
types, which are known, see \code{help(countsSorted2)}, were renamed in the
spUnsupervised object. The \code{permuteSwarm} was then run and both the
spSwarm object and the output from the \code{permuteSwarm} function were
saved.
}
\examples{
data(sObjPermutations)
}
\keyword{datasets}
