% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabula.muris.data_TMD.R
\docType{data}
\name{TMD.Counts}
\alias{TMD.Counts}
\title{Counts data from mouse colon in Tabula Muris study.}
\format{Matrix containing counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
TMD.Counts
}
\description{
Counts, ERCC, and metadata from mouse colon in Tabula Muris study.
}
\examples{
data(TMD.Counts)

}
\keyword{datasets}
