% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabula.muris.data_TMD.R
\docType{data}
\name{TMD.CountsERCC}
\alias{TMD.CountsERCC}
\title{ERCC counts data from mouse colon in Tabula Muris study.}
\format{Matrix containing ERCC counts with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
TMD.CountsERCC
}
\description{
Counts, ERCC, and metadata from mouse colon in Tabula Muris study.
}
\examples{
data(TMD.CountsERCC)

}
\keyword{datasets}
