% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabula.muris.data_TMD.R
\docType{data}
\name{TMD.Meta}
\alias{TMD.Meta}
\title{Metadata from mouse colon in Tabula Muris study.}
\format{Tibble}
\usage{
TMD.Meta
}
\description{
Counts, ERCC, and metadata from mouse colon in Tabula Muris study.
}
\examples{
data(TMD.Meta)

}
\keyword{datasets}
