% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_general.R
\name{writeMetaData}
\alias{writeMetaData}
\title{Write meta data.}
\arguments{
\item{data}{data.frame or tibble; the metadata to be output.}

\item{path}{character; The path to the output metadata file including the
file name. For excel, typically in a subfolder of inst/rawData. For Rdata,
typically in the data directory.}

\item{type}{character; the type of file to write. Can be either "excel" or
"Rdata".}
}
\value{
The metadata as a tibble.
}
\description{
Writes a metadata as an excel file or .rda file.
}
\author{
Jason Serviss
}
