% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetalPancreasCounts.R
\docType{data}
\name{countsFp}
\alias{countsFp}
\title{Fetal pancreas counts data.}
\format{matrix
\describe{
    \item{rownames}{Gene names. ERCC identified with "^ERCC-[0-9]*"}
    \item{colnames}{Samples. Singlets prefixed with "s" and multiplets "m".}
}}
\usage{
data(countsFp)
}
\description{
Fetal pancreas counts and ERCC data.
}
\examples{
data(countsFp)
}
\keyword{datasets}
