% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFuns.R
\name{labelSingletsAndMultiplets}
\alias{labelSingletsAndMultiplets}
\title{labelSingletsAndMultiplets}
\arguments{
\item{counts}{data.frame; A data frame with counts data.}

\item{ids}{character; a character vector of regex statments used to indicate
colnames of columns that contain SINGLETS.}
}
\value{
The counts data.frame with the appropriate prefix attatched to the
 column names. Throws a warning if no singlets are detected.
}
\description{
Adds the prefix "s." to colnames of columns that contain singlets and "m." to
columns that contain multiplets.
}
\examples{

counts <- data.frame(a = runif(26), b = runif(26), c = runif(26))
labelSingletsAndMultiplets(counts, ids = c("a", "b"))

}
\author{
Jason Serviss
}
