% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\docType{data}
\name{sObjPermutations}
\alias{sObjPermutations}
\title{Permutations of the countsSorted2 dataset.}
\format{An spSwarm object containing the "real" results.}
\value{
The spSwarm object containing the "real" data.
}
\description{
The countsSorted2 dataset was used to run the sp.scRNAseq pipeline. The cell
types, which are known, see \code{help(countsSorted2)}, were renamed in the
spUnsupervised object. The \code{permuteSwarm} was then run and both the
spSwarm object and the output from the \code{permuteSwarm} function were
saved.
}
\examples{
data(sObjPermutations)
}
